Added missing cast

+++ Core/FileSystems/FSDescriptor.cpp
@@ -124,7 +124,7 @@ FSDescriptor::checkCompatibility() const
     if (!FSFormatEnum::isValid(dos) || FORCE_FS_WRONG_DOS_TYPE) {
         throw AppError(Fault::FS_WRONG_DOS_TYPE);
     }
-    if (rootBlock >= numBlocks) {
+    if (isize(rootBlock) >= numBlocks) {
         throw AppError(Fault::FS_OUT_OF_RANGE);
     }
 }
